test(hooks): split oversized LSP guard test files below 500 lines (ADR-062)
Move entry-point/bootstrap classes (read guard) and Main/subprocess classes (pre-delegation guard) into sibling test modules to clear the file-size taste-lint. 96 tests across the 4 files; total suite unchanged at 1314.
